fix: rename template file
