+++ core/src/ten_manager/designer_frontend/src/App.tsx
@@ -28,7 +28,7 @@ import { GlobalDialogs } from "@/components/GlobalDialogs";
 import Dock from "@/components/Dock";
 import { useWidgetStore, useFlowStore, useAppStore } from "@/store";
 import { EWidgetDisplayType } from "@/types/widgets";
-import { GlobalPopups } from "@/components/Popup/GlobalPopups";
+import { GlobalPopups } from "@/components/Popup/Global";
 import { BackstageWidgets } from "@/components/Widget/BackstageWidgets";
 import { cn } from "@/lib/utils";
 import { usePreferences } from "@/api/services/common";
