+++ core/src/ten_manager/designer_frontend/src/components/Popup/Default/Preferences.tsx
@@ -10,7 +10,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 
-import { PopupBase } from "@/components/Popup";
+import { PopupBase } from "@/components/Popup/Base";
 import { Label } from "@/components/ui/Label";
 import { Button } from "@/components/ui/Button";
 import {
