+++ core/src/ten_manager/designer_frontend/src/components/Popup/Graph.tsx
@@ -7,7 +7,7 @@
 import * as React from "react";
 import { useTranslation } from "react-i18next";
 
-import { PopupBase } from "@/components/Popup";
+import { PopupBase } from "@/components/Popup/Base";
 import { useWidgetStore } from "@/store/widget";
 import {
   GraphAddConnectionWidget,
