+++ core/src/ten_manager/designer_frontend/src/components/Popup/LogViewer.tsx
@@ -8,7 +8,7 @@ import * as React from "react";
 import { useTranslation } from "react-i18next";
 import { PinIcon, OctagonXIcon } from "lucide-react";
 
-import { PopupBase } from "@/components/Popup";
+import { PopupBase } from "@/components/Popup/Base";
 import { LogViewerFrontStageWidget } from "@/components/Widget/LogViewerWidget";
 import {
   EWidgetCategory,
